Bluetooth key missing handling for LE transport
When we connect to peripheral we were bonded with, and it claims it is
missing keys, user must go to settings, and forget this peripheral
before bonding with it again.

Some users donn't know what they should do, or they just observe their
device stopped working and don't get any hint what to do.

After this aptch, Bluetooth stack will send a Broadcast that can be
handled by Settings or Fast Pair, and acted upon properly.

Test: Bond with device, remove bond just from peripheral, try to reconnect, observe dialog
Bug: 297598060
Bug: 311447399
Change-Id: I5ddde4da2c66280a9408fad2f960ffc352f82ee3