LeAudio: Improve LTV map structure
Add the compare operators, a way to get two maps intersection
and a more convenient raw data parsing helper.

This is not used with the current set of features but will be
needed for the upcoming multi-codec support.

Bug: 295972694
Test: atest --host bluetooth_le_audio_test bluetooth_le_audio_client_test bluetooth_test_broadcaster bluetooth_test_broadcaster_state_machine bluetooth_le_audio_codec_manager_test
Flag: EXEMPT; not used yet (will be used by the upcoming multicodec feature), verified with unit tests
Change-Id: Ia65e480497121e624ff87d471b65d325e7010528