floss: hcidoc: Remove associated CID when L2CAP is disconnected
It's a good bookkeeping practice to mark CID as unassigned when L2CAP
is disconnected. Furthermore, there could be a race from both sides of
the connection to disconnect the same CIDs. If we free those CIDs on
the first response, we won't process duplicate disconnection reports.

Bug: 293397988
Test: m -j
Flag: EXEMPT floss only changes
Change-Id: If0de7b21ffc8d998753077042bed0073e7e59df7