Don't add the contents of dump() to our log messages
We're only interested in the device map, yet we execute dump() and add
it to a few logs to get very verbose debugging info. This isn't
necessary. We should print what we need, and use dumpsys for more
logging if we need it.

Tag: #refactor
Flag: EXEMPT, debug output change only
Bug: 319696083
Bug: 315046089
Test: atest MapClientServiceTest.java
Change-Id: I882438a2ce28bf67dc3b077f26c4f12563cf3a9d