leaudio: Fix group IsInTransition
This should return true only when group is truly in transition, meaning
Target state was set by the stack and this is different from current
state.

Group is not in the transition, when after restart ASEs are autonomously
set to configured state.

Bug: 318709435
Test: atest bluetooth_le_audio_test
Tag: #feature
Flag: Exempt, LeAudio fix verified with unit tests. Also new test added.

Change-Id: I2b07283e3be24ab7e7512a007152c385be5a9743