Fix SMP total_tx_unacked counting
total_tx_unacked needs to be incremented before L2CA_SendFixedChnlData
is called, because it is possible to call smp_tx_complete_callback
immediately, in which we should decrement the value but we can't since
it is still zero.

If this happens, we will be stuck with total_tx_unacked = 1 after the
SMP transaction is done, and we don't actually detect it as complete.

Bug: 319027651
Bug: 314819704
Test: m -j
Test: Pair with LE device, and run LE pairing tests.
Change-Id: I1fd32297f77f5d61ccd80f41f87639c1b3b3df56