{AvrcpTarget,HeadsetClient}Service: Throw Exception on double start
This aligns the behavior with other services that checks for
double starts like: A2dpService, HeadsetService, LeAudioService,
BatteryService, HearingAidService, ...

This is done to remove the boolean return of the start method
in a child change.

This has a potential to create a regression, but is unlikely
as those profiles are started similarly than other profiles
containing the same check.

Bug: 319112725
Test: m Bluetooth
Flag: Exempt, Accelerate release of this change to be able to
  perform a mechanical refactor on top of it.
Change-Id: I0b611622525f7a741d0962b1d1a827bd56ba1a3f