ProfileService: Change `stop` return to void
The return of ProfileService#stop was only used to log an error.
And all stop implementation also log a line before returning false.

So the return is changed from boolean to void to be able to remove
all checks of ProfileService#stop's return value and simplify
the callers.

Bug: 319112725
Test: m Bluetooth BluetoothInstrumentationTests
Flag: Exempt, mechanical refactor
Change-Id: I6df7562859ced79db86d7900bd7499ca65165c22