Avoid active scanning if device is discovered but not connected
If the device is connected but not yet added to group since sirk exchange
is not completed, we cannot start active scan because this device would
be in the group eventually. Open the second active scan is not necessary
if all members in the group are found and being processed.

Test: atest -c bluetooth_csis_test
Bug: 315392056
Tag: feature
Flag: Exempt, regression covered with unit tests, new unit test added
Change-Id: Ibbb0a8e98bcf475b4f294b1fbba6bf135606c23a