csis: Fix notification of device available on startup
It might happen that native Csis module will start later than Group module and
in this case `onDeviceAvailable` callback will not be called, as at this
point of time, Rank is not known.
Rank being not available at this point is fine as after connection Rank
will be updated.

This might result in invalid device represantation in GUI as Java is not
aware about grouping.

Note, with this patch behavior will match to the one when native Group
module starts before Csis module.

Bug: 312997962
Test: atest bluetooth_csis_test
Flags: Exempt, simple fix, regression verified with unit tests
Change-Id: I69e50e5e046730b6dcf51e12d008f1064f4e8e8c