Fix BluetoothMapSettingsTest failure in some devices
In some devices, the BluetoothMapSettings activity doesn't fill the
entire screen, and also don't get focus when they pop up.
As a result, Espresso uses a wrong root view to find the desired view
in, which makes the test fail.

This CL adds a click event on the activity so that the correct
root view can be used for finding the desired view in it.

Also it adds @FlakyTest since it is flaky due to intrinsic nature
of UI tests. Among 50 runs on different devices:
- Did not fail on a phone device.
- Failed about 3% in a relatively slow device.

Bug: 322817893
Test: atest BluetoothMapSettingsTest --iteration 50
Flag: EXEMPT, test-only change
Change-Id: I7a83aeb538a2a863da01e4a4ae0688d1331e9e16