Drop codec configuration changes for disconnected devices
Currently, if a device associated with a config changed event isn't
null, we look for it in the state machine map. If the device is
disconnected, the state machine has been cleaned up and returns a null
instead. We use the null state machine, which gives us an NPE.

For now, we can drop the configuration changed event for the null
device so we can avoid a crash that would have otherwise logically
dropped the event anyways.

Tag: #stability
Bug: 324073867
Flag: EXEMPT, null check to avoid a crash + test coverage
Test: atest A2dpSinkServiceTest#testOnAudioConfigChanged_withUnknownDevice_eventDropped
Change-Id: I8531ae5697404a214254be13109d9ec290903fbb