leaudio: Fix for VOIP apps not using Telecom
If VOIP  application is not using Telecom then Bluetooth is blind when
it comes to making sure that it is in a Call scenario.
It can only based on the received context types of the stream and it
turns out that when calling, some applications are using:

usage=USAGE_VOICE_COMMUNICATION_SIGNALLING(3), content_type= CONTENT_TYPE_SONIFICATION

which currently is mapped to MEDIA. It make it inconsistent with other
VOIP and Phone app. To make it work, Bluetooth need to map this setup to
Ringtone. This patch fixes that.

Bug: 307873884
Flag: Exempt, trivial fix for some VOIP apps
Test: mmm packages/modules/Bluetooth
Test: manual verification of the VOIP call
Change-Id: Iabf5c35590844757ea7dfc11b8d267ebfdd41c0c