Fix ScanManagerTest#testDowngradeDuringScanForConcurrencyScreenOff
The test is flaky due to MSG_STOP_CONNECTING not handled before
assertions of scan mode changes, since MSG_STOP_CONNECTING is sent with
delay and the current test is just waiting for the delay to end with
sleeping. This fix uses the testLooper to change the test to assure that
MSG_STOP_CONNECTING is always handled before test assertions by using a
configuring the handler clock.

Bug: 324806849
Test: atest BluetoothInstrumentationTests
Flag: EXEMPT, test fix only
Change-Id: I9299cae6689e9dc0670bcdbe31f0d053431f5e19