LeAudio: Remove the invalid audio set configurations
Four ASEs configurations are not supported by BAP 1.0.1 and not
supported by the existing implementation. The stack will always
choose the the target strategy for group with more than one device,
as the MONO_ONE_CIS_PER_DEVICE strategy. These configurations will
never be selected as they are defined with 2 devices and the
STEREO_TWO_CISES_PER_DEVICE strategy.

Two ASEs per device configurations are only supported with
a single device in the group (banded headphones).

This is part of the cleanup and refactor phase for the multi-codec
support.

Bug: 308427707
Test: atest bluetooth_le_audio_codec_manager_test
Flag: EXEMPT; Not flaggable change, verified with unit tests
Change-Id: Ie4fd010aeda14332ea325217212df4f4a21babd1