floss: Fix edge case when adv started after client has disconnected
If a client disconnected right after StartAdvertisingSet before
OnAdvertisingSetStarted, the advertiser could be leaked and never
cleaned up.

This patch marks the not-yet-valid advertising sets as stopped and
remove them on they started.

Since the function now lives in bluetooth_adv.rs and has the full access
to the internal structures, we could get rid of the unnecessary clone.

Bug: 242083290
Tag: #floss
Test: mmm packages/modules/Bluetooth
Test: Manual test advertising with NearbyShare
Flag: EXEMPT, Floss-only changes
Change-Id: I7b2b0837d886b3160b43ce3b61500a06099fbce8