[le audio] Set pa sync interval default to match the APIs
getPaSyncInterval() and setPaSyncInterval() both declared the unknown
interval value is 0xFFFF.
But in metadata builder, we set the default as -1.
In the case of assistant parsing metadata from Base, pa interval is not
present and we uses -1 there which is not matching the APIs.
We should update the default value to 0xFFFF to keep aligned with APIs.

Tag: #bug
Bug: 326528942
Bug: 316005152
Test: atest BluetoothLeBroadcastTest BluetoothLeBroadcastMetadataTest BluetoothLeBroadcastAssistantTest
Change-Id: I9adf0356cbf0f52907ca816cb3aae84a4f58dd6b