Fix crash by passing original slot id to avoid race condition.
There seems to be a race condition where the slot is getting destroyed and a 0 slot id keeps getting passed.
This fix eliminates that issue. We also remove the `CHECK` since it introduces the crash and a slot id == 0 is a recoverable state.

Fix: 316684925
Bug: 316684925
Test: m .
Flag: EXEMPT Eliminates crash with a simple change. No functional change.
Change-Id: I9f29423683ce1d4dff2bddacc87b45ee668d1db4