Fast bind to app implementation
OnCreate method is intended to return almost immediately in order for
the system Server to be notified in case of a crash.
By moving the initialization from OnCreate to "enable" we can be
notified of any crash happening during initialization and restart it a
few second later (instead of 90 min later by the system).

Bug: 328698375
Bug: 323672160
Test: atest CtsBluetoothTestCase
Test: atest pts-bot
Test: atest ServiceBluetoothTest
Change-Id: I34da95f878879c14e26b3989e676f71617a324ec