Cap total no. of max connected devices
Problem: Bluetooth stack can support upto total 5(configurable)
audio devices. This includes both A2DP source and sink devices.
Currently, each profile accounts for the devices connected with
its profile and doesn't account for the no. of devices connected
in other profiles.

Solution: While checking if a device is allowed to connect or not,
it will also account for the other profile's no. of devices
connected.

Bug: 312533591
Bug: 321806163
Test: atest net_*
Change-Id: I59666577f615dda3c8b0fa2a9b1246cf645c0c6d