interop: allow hid host to connect to not-first HOG service
Due to the limitations of the current HID host implementation, it
assumes that the HID peripheral has only one HOG service, or if there
are multiple HOG services, the first one is always active. To address
this issue temporarily, an interop quirk has been introduced to allow
the HID host to connect to a non-first HID service for certain devices.
However, for a long-term solution, the HID stack needs to be refactored
to handle multiple HID services simultaneously.

Bug: 286413526
Bug: 328129633
Tag: #floss
Test: mmm packages/modules/Bluetooth
Test: Manual test on Brya, Logitech CASA working
Flag: EXEMPT, apply only to one and not-functioning device

Change-Id: I0733f1597d9058c992e46b99c6a31748714d8dfa