leaudio: Fix crash on metadata update
When moving metadata to Bluetooth mainthread, make sure that Bluetooth
gets metadata and not just a pointer to memory on the stack.

`sink_metadata_v7` contained a pointer to array of either (record|playback)_track_metadata_v7,
so we can't rely on simple struct copy constructor. Pass the array content as vector instead.

Bug: 329067188
Test: mmm packages/modules/Bluetooth
Test: manual LeAudio streaming with different contexs (including call)
Test: atest bluetooth_le_audio_client_test
Ignore-AOSP-First: security
Change-Id: I9c01c81c0094df87efc1b3885ad69d82b0970404