HidHost: Use correct device instance for opening uhid driver
Previously, reconnect_allowed field was added in the device instance to
keep track of connection policy. This required that the device instance
is created at the time of connection request and also when the devices
are loaded from the persistent storage. This meant that device instances
for the previously bonded HID devices were created at the time of
initialization. When corresponding devices reconnected, it was possible
that the wrong device instance might get selected. Which also meant that
the wrong connection policy may get applied.
This new change moves the reconnect_allowed field to the added device
instance. As a result, device instances are not needed to be created at
the time of intialization and there is no possibility of applying wrong
connection policy.

Test: mmm packages/modules/Bluetooth
Test: Manual | Connect two HID devices, disable the first device,
restart Bluetooth, initiate connection from the second HID device.
Bug: 330514181
Bug: 320762367

Change-Id: I0c5f61613d9b159939f4ae1566a1c417d45119c2