Fix MAP Server test with NPE when debugging is enabled
This test didn't actually run the code path it thought it did. When the
code path is executed it has a NPE because the cursor isn't actually
initialized to point at an element.

This change updates the test to move the cursor, and adds some simple
defensive programming changes to the actual log function under test so
it shouldn't crash in the future if a similar problem arises. I also
updated the logging pattern to match new guidelines.

Tag: #stability
Flag: EXEMPT, test fix + logging only function
Bug: 315046089
Test: atest BluetoothInstrumentationTests
Change-Id: I3f306996aff584da12b439924006b279b91dcfb2