Floss: RTL8852 MSFT support
Because RTL8852 supports only one peer for each monitor, it needs a
special handling to make pattern searching via MSFT work.

When a peer device with the matching pattern is scanned, we need to
split that into its own address monitor. At the same time, we also
report device found. This way we can monitor multiple peer devices.
When the address monitor reports device lost, we cancel the address
monitor and reports device lost to the host as well.

Bug: 294961421
Bug: 332438857
Tag: #floss
Test: bluetooth_AdapterAdvMonitor.all_floss
Test: m -j

Change-Id: I2d5679e5395fd04805fc2d83dd73df7b295af7cc
Signed-off-by: klay_liu <klay_liu@realsil.com.cn>