Framework: Add the flag library in stub
Since this is adding a shared library and not a static, we expect this
to not have any side effect.
But this is a hack to workaround metalava that parse unconditionally
some javadoc. The Flags library should be considered hidden (or
assimilate) and not be exported in metalava

Bug: 332362688
Test: m the correct target (…/framework-bluetooth.stubs.from-source.jar)
Test: m out/soong/.intermediates/packages/modules/Bluetooth/framework/framework-bluetooth.stubs.exportable.from-source/android_common/javac/framework-bluetooth.stubs.exportable.from-source.jar

Flag: Exempt, comment only
Change-Id: I795b609b5a05829bd4f00a43989daea34290e7c3