LeAudio/Brodcaster: Adjust the configuration struct
This defines the channel count explicitly in the static configuration
structs for the broadcast codec configurations, instead of deriving it
from one of the LTV parameters.

This is to match the AIDL config provider and decouple from the
LTV allocation counter channel count, which is not vendor codec
friendly (uses LTV list).

Bug: 308428860
Test: atest bluetooth_le_audio_test bluetooth_le_audio_client_test  bluetooth_test_broadcaster bluetooth_test_broadcaster_state_machine bluetooth_le_audio_codec_manager_test
Flag: EXEMPT; no functional change, parameter extraction veirfied by the unit tests
Change-Id: I95fb89191717f4f560fdc89e11d99d34ce4ea39f