Gatt Rust: Convince rustc that OwnedHandle is used
OwnedHandle is dropped when a connection is dropped with a
Pending transaction.  Removing this will change the behavior.

warning: field `0` is never read
  --> src/gatt/server/att_server_bearer.rs:37:13
   |
37 |     Pending(OwnedHandle<()>),
   |     ------- ^^^^^^^^^^^^^^^
   |     |
   |     field in this variant
   |
   = note: `#[warn(dead_code)]` on by default
help: consider changing the field to be of unit type to suppress this warning while preserving the field numbering, or remove the field
   |
37 |     Pending(()),
   |

Bug: 330185853
Test: atest libbluetooth_core_rs_test
Flag: EXEMPT, build-only change
Change-Id: I00d61e2a1f91667931edf0beaea7961c5e00296b