leaudio: Fix reconfiguration for sonification types
There is a logic which prevents reconfiguration for the sonification
contexts. This is usefull especially after GAME or CONVERSATIONAL
stream, when in the end of stream there is either SOUNDEFFECT or
similar. In such a case we just stream that on previous configured
CISes.

However, there is a bug, that if CISes are not there, we keep using old
configurations for sonifications which could lead to send NOTIFICATIONS
on bidirectional CISes if NOTIFICATION is send after phone call.

This patch fixes that

Bug: 317660246
Test: atest bluetooth_le_audio_client_test
Flag: Exempt, regression tested with unit test and new test added to
cover error scenario

Change-Id: Ied7e09414edfbbeb3977d1898e034e77a886dc23