LeAudioService: Fix behaviour of getCodecStatus
With this patch we make sure that if INVALID_CODEC codec is provided by
native it is not on the SelectableCodecConfig list and also not used as
and mInputCodecConfig and mOutputCodecConfig.

With this patch if e.g. for current configuration does not support
input, then getCodecStatus.getInputCodecConfig() will return null.

If user wants to know if given direction is supported by the remote
device, he can use getInputCodecSelectableCapabilities() and
getOutputCodecSelectableCapabilities(), and if it is non empty, it means
direction is supported

Bug: 314094343
Test: atest LeAudioServiceTest
Flag: Exempt, regressions test with unit tests, new test added
Change-Id: Iee754b51779d778d1b8e557e0698ea8d385f1372