Remove test that triggers undefined behavior
According to C++ standard, compiler is allowed to produce arbitrary
result if undefined behavior is triggered. It is not meaningful to have
a test to assert the behavior.

We tried to let UBSAN ignore the violation, but it does not work because
the UB happens in a different shared library.

Test: atest bluetooth-test-audio-hal-a2dp-provider-info64.ProviderInfoTest.TestCodecIndexStr
Bug: 334078285
Change-Id: I21bc60fd3211670df01538677b3147e07ba1042e