flags: uncache_player_when_browsed_player_changes
Introduce a flag that to fix a bug where we don't uncache a player
object when we switch to a new player. This keeps us from ever being
able to get fresh player contents when things get updated and we haven't
received an available players changed event. This also doesn't fit in
with our current caching scheme.

Bug: 337259487
Bug: 337257397
Bug: 315241296
Test: m com.android.btservices
Change-Id: I6f36be47a3578a163cc67bcde8348e3c4098c22d