Null check and print error messages for failed MMS/SMS DB interactions
MAP Client requires the Telephony MMS/SMS DB to be up to store messages.
It should always be up when the feature is declared, but we still want
to avoid crashes and have useful error messages when it is not.

A longer term refactor to prevent messaging until the DB is up (or
whatever we decide the right thing is) should come in the future so we
can be even more robust to this.

Bug: 339911228
Flag: EXEMPT, null check + logging changes only, with tests
Test: atest com.android.bluetooth.mapclient.MapClientContentTest
Change-Id: I72490da32ae4429548170a7a04e75ddc3e18dfd1