floss: Remove Uuid128Bit
It makes no sense to keep both Uuid and Uuid128Bit in the Rust space as
Uuid is simply a wrapped Uuid128Bit.
Keep Uuid but not Uuid128Bit because Uuid128Bit is a builtin type
and we can't impl any helper methods for it.

Bug: 342337056
Tag: #floss
Test: mmm packages/modules/Bluetooth
Test: bluetooth_AdapterQuickHealth.AVL.all_floss
Test: NearbyShare / PhoneHub / SmartLock / FastPair
Flag: EXEMPT, Floss-only changes
Change-Id: Iea9ae139c8afad25a8e919ffc0d4f23c5de87cf8