floss: propagate VC connection event to audio server
The audio server needs this information to determine whether software
volume adjustment is needed.

If VCP is connected before being identified as a group, the callback
is delayed until the group is notified to the audio server. This makes
things much simpler given the assumption that a device attaches to a
unqiue group id in the lifetime of its connection.

Bug: 317682584
Test: m Bluetooth

Change-Id: I45fa4309281fb22cc68651763f492bb48b539462