hid: separate the struct accessed by the uhid thread
The uhid thread shared the same resource with btif and main thread,
making it easier to crash. This can be prevented by making sure the
uhid thread has an exclusive access to a part of the memory that is
not controlled by the other threads.

This CL splits the information needed by the uhid thread into its own
struct, btif_hh_uhid_t. However, this is only a mechanical refactor,
the btif thread can still access this part of memory like before.
A future CL shall make sure that uhid thread fully manages the struct
btif_hh_uhid_t and has exclusive access to it.

Bug: 291522341
Test: mmm packages/modules/Bluetooth
Flag: EXEMPT, mechanical refactor

Change-Id: Ied37b336a0d6b36a2b1c438381c6388bfce13c1b