LeAudio: Fix configuring incomplete group
When codec extensibility feature is enabled we get the whole group
configuration even if just a single device is currently connected.
When configuring a single device, the ASE configurations should be
filtered based on the audio allocation, to avoid configuring two
ASEs on a single device (with just one audio location).

Bug: 308428644
Bug: 313054645
Test: atest bluetooth_le_audio_test bluetooth_le_audio_client_test
Change-Id: I2d5a6a3a5f47915bf92098841a415aa048d1d64d