hid: fix assigning empty handle to uhid
aosp/3119833 introduces a bug when separating uhid and hh struct.
When the hh struct was just created, the handle is still unassigned,
but is nevertheless passed to the uhid struct. Likewise for the
link_spec member.

This CL makes sure to pass the correct values to the uhid struct.

Bug: 291522341
Test: mmm packages/modules/Bluetooth
Test: manual verification that set/get feature reports are forwarded
      to the correct peripherals.
Flag: EXEMPT, bugfix to previous mechanical refactor
Change-Id: I058429446595b687ddfba2dd9693f4c97acc3f85