a2dp: use peer capability instead of config
In case of incoming connection, is_capabilty comes as
false in setCodecConfig, which leads to ota_codec_peer_config_
to be updated to as that configured by remote.
But when channel mode is changed from Dev UI, the same
config, which was configured by remote is used instead of config
set from DUT.

In the fix, removing the part where peer config is being used,
and use peer capability only.

This behavior was previously introduced by
https://r.android.com/395892 as workaronud to pass selected
pts tests; it does not seem to be necessary anymore.

Bug: 395551185
Flag: EXEMPT, minor bug fix
Test: atest pts-bot
Change-Id: Ia4e18a7a74b8c403a904b4db30d16004a2205fca