VCS: Fix deadlock on handleStackEvent
A deadlock is caused by synchronization on a method (this) and
mStateMachines. In many synchronized method is used "synchronized
(mStateMachines)" which is fine till we not call such methods from
"synchronized (mStateMachines)" scope.
Such situation has place on first event after bud connection
EVENT_TYPE_DEVICE_AVAILABLE when StateMachine is not yet available.

With 4 and more buds, it is very easy to reproduce the deadlock by
simple turn off and on Bluetooth to force simultaneous connection
of all of them.

With that fix it does not happen (tested with 6 buds).

Bug: 396599136
Flag: EXEMPT, day-to-day fix covered by unit and manual tests
Test: atest VolumeControlServiceTest
Change-Id: Id013d88382db01c4ae79effcc76f76b20edb8682