Simplify LRU cache mechanism for remote device properties
It is not obvious that the mDeviceQueue is used to make mDevices an
LRU cache. This has led to bugs in the past. We are now simplifying
the logic by removing mDeviceQueue and making the mDevices a
LinkedHashMap.
Also if the eldest device map entry is a bonded device, adding new
device does not eject any entry. This results in size of the device
map going beyond the limit of 200 devices. This change now ensures
that the size of the device map does not exceed the stated limit.

Test: mmm packages/modules/Bluetooth
Flag: EXEMPT bugfix
Bug: 403392250

Change-Id: I62c8c7114000983e9ff99c6f99ab5b73676e8d54