Refactor A2DP disconnection test in PairingTest
Move the A2DP disconnection code to a separate function because as soon as the devices are bonded, profile connections are initiated and if the Intent registeration is delayed, it caused issues with in-order verification.
To avoid this, replaced the state transitions with the state matching with a delay and refactored the current test into individual testSteps.

Test: atest PairingTest
Flag: TEST_ONLY
Bug: 398746370
Bug: 378268278
Change-Id: I041f07f04358668e9e67ff7c4661c8ab6efe15e2