Revoke permissions on rebond attempt
When a device with the same address
as a previously bonded device attempts
to bond again, the adapter service may not
receive a BOND_NONE state change message,
even if the new device is different.

This can prevent the user from being prompted
to grant consent for the new device.

This commit ensures that permissions are revoked
when a rebond atempt is made with the same address,
ensuring the user is prompted for consent with
the new device.

Bug: 405333875
Bug: 290364858
Flag: com.android.bluetooth.flags.reboke_permission_on_unbond
Test: atest AdapterServiceTest
Change-Id: Id63a61908fb22c426c49d0979bba0c6a8b9dba03
(cherry picked from commit 921aaec06df0b5974d0fbbb40443935b15019270)