Rename `startScanForIntent()` to `registerPiAndStartScan()` for Binder consistency
Initially, I thought `startScanForIntent()` was a better name but looking at the implementation
in `ScanController#registerPiAndStartScan()`, the function does not technically start a scan in it.
It calls `mScanManager.registerScanner()` which later arrives at `onScannerRegistered()` which
later starts the scan in `continuePiStartScan()` via `mScanManager.startScan()`.
So the name `registerPiAndStartScan()` seems more accurate of the actual process.

Follow up ag/32575738/comment/455c563c_e97fc053/

I was not able to rename to `startScan()` due to overloading
not being supported in AIDL b/258604183, b/184894712.

Bug: 405227923
Test: atest ScanBinderTest
Test: atest ScanControllerTest
Flag: EXEMPT refactor no-op
Change-Id: I58eb0b32d2fe80ff9d809a041ba33469f3796945