leaudio: Clear RINGTONE context type on call end
When call ends, a speedup reconfiguration happens so the Media can
start as soon as possible.
However, when Ringtone context left, the reconfiguration will not happen
as stack will assume VOIP scenario.
This patch fixes that.

It also fixes the test case which was testing UNSPECIFIED metadata being
used for Unsupported context type on existing stream. So far it worked
only because mentioned RINGTONE keep stream running after call was ended
and speed up reconfiguration was not actually working.

Bug: 406694283
Test: atest bletooth_le_audio_client_test
Flag: EXEMPT, day to day fix, regression covered with unit tests
Change-Id: I3c662eab71f84ee136f428c0d2a4a6a217f5d7b4