Don't remove device properties on disconnection of non-bonded devices
Device properties are removed when the device is unbonded and ACLs are disconnected. However this logic also removes the device properties of devices which were never bonded on ACL disconnection.
This means if an app, for example, connects and disconnects with a recently discovered device without pairing with it, the device properties like device name would not be available unless device discovery (or LE scanning) is performed again.
This happens because there is no way to distinguish between the devices which never bonded and devices for which bond has been removed.

Test: mmm packages/modules/Bluetooth
Flag: com.android.bluetooth.flags.non_bonded_device_properties
Bug: 406601059
Bug: 406602074
Change-Id: Ia914177e7590ce2207b901d11caf282693622f07