leaudio: Fix for missconfigured sink only device
We tested against the device which was having Source PACS and Source
Context types and in the same time it was missing ASE Source.

Such missconfigured device trigger issue on our end with a
CONVERSATIONAL context type, as Android is trying to open MIC on this
device which is impossible.

The root cause was that native stack incorrectly provided to Java
supported directions which lead to open IN and OUT in the Audio
Framework.
This patch fixes that.

Bug: 400579870
Test: atest bluetooth_le_audio_client_test
Flag: EXEMPT, day to day fix, regression tested with unit tests
Change-Id: I8a3bdc5e0a35aa647c940911339f2a3c752b3426