Set correct address type in the device properties
Apps can set the address type when creating BluetoothDevice objects. But when created by the stack, the address type is never set in the BluetoothDevice objects. This means that all BluetoothDevice objects created by the stack would appear to have a public address to the applications.
This change now ensures that the address type is set in the BluetoothDevice objects when stored in the cache.

Test: mmm packages/modules/Bluetooth
Flag: com.android.bluetooth.flags.retain_address_type
Bug: 358566084
Bug: 407277344
Change-Id: Ie5c97f16677ac85f7144febd72fbfb364a41c8c2