system/gd: Fix ordering issues
I found that running the tests in a random order caused some of
the snoop logger tests to fail. To Fix this, I have changed the tests
so that the failing tests bind to an ephemeral port and then it seems
they can run in a random order without failing.

Bug: 407647836
Fix: 407787521
Flag: TEST_ONLY
Test: atest bluetooth_test_gd_unit
Change-Id: Ia6d728811fd81dd469dd73f3e16fef075d776468