Bumble: Make asynchronous gRPC calls be cancelable
(This is a re-upload of 5bd7789 with build fix applied.)

When getting responses from gRPC calls using ResponseObserver, there
were no way to cancel the started async calls.

This lead to the failure when we do BLE scan two times in a row, since
the previous scan is not canceled.

This CL makes StreamObserverSpliterator implement ClientResponseObserver
so that it can get ClientCallStreamObserver instance which has cancel()
method.

Bug: 408420093
Flag: TEST_ONLY
Test: atest BumbleBluetoothTests
Test: atest LeAdvertisingTest#advertisingSet (w/o @Ignore)
      Canceling previous scan made another scan be started w/o error.
Change-Id: I7ccc8ab75c165b512d8e79eaa37a7fa442232c99