Suppress redundant and incompatible checkstyle
google-java-format is source of truth since it provide a fixer.
For anything that it can do, we should not have a contradictory message
from checkstyle.
Indent is a disagreement in switch case with braces.
LineLength is a disagreement when using very very long variable (like in
metrics).

Test: m .
Bug: 311772251
Flag: EXEMPT dev feelgood
Change-Id: Id4723ff49d1c3f7ff6734fccfcca1e5202d8b032