Increase TIMEOUT_MS to 4000 for DckScanTest and DckTestRule
This also un-ignores tests within DckScanTest

This is a pretty safe change. The timeout given here is a maximum
window allowed for the tests to enable / disable bluetooth and
do the required operation such as advertise and scan. Increasing
the timeout does not slow down the test in majority of cases and
only allows additional time for slow cases to recover and not fail.
In cases where it succeeds within 1-3 seconds, as the average test
duration, this change does NOT slow it down.

A similar fix was implemented for LeScanningTest in aosp/3453632

Bug: 406277877
Bug: 404159990
Fix: 406277877
Fix: 404159990
Test: atest DckScanTest
Flag: EXEMPT test only
Change-Id: Id90f0bc506b142b76c12bca34c18b4cf020ee7c7