Increase TIMEOUT_MS to 5000 for DckScanTest and DckTestRule
The framework sometimes takes longer to
start/stop bluetooth and start/finish scanning. Increase the timeout to reduce test flakiness.

This is a pretty safe change. The timeout given here is a maximum
window allowed for the tests to enable / disable bluetooth and
do the required operation such as advertise and scan. Increasing
the timeout does not slow down the test in majority of cases and
only allows additional time for slow cases to recover and not fail.
In cases where it succeeds within 1-3 seconds, as the average test
duration, this change does NOT slow it down.

A timeout of 4000 was implemented in ag/32895870

A similar fix was implemented for LeScanningTest in aosp/3453632

Bug: 406277877
Bug: 404159990
Test: atest DckScanTest
Flag: EXEMPT test only
Change-Id: Ie4eb7516c392fcd32b53c3f3573fc9fdf08767b1