audio_hal_interface/a2dp: Make all FMQ reads exact
Reading from the FMQ must return the exact number
of samples requested, or 0. There is logic
implemented in all software encoders to handle
incomplete reads, but most of the time the audio
buffer is padded with zeros. This is very wrong
as it generates audio blips and alters the stream
latency.

Bug: 406437701
Bug: 409124193
Flag: com.android.bluetooth.flags.a2dp_fmq_read_exact
Test: m com.android.bt
Change-Id: I85bb3de4c2b5357f7b7ee242e2271400eb2c996c