ahal: remove supports_codec dependency in stack
When the codecs are initialized on the startup, it is checked whether
hal provider supports the codec or not - based on that the codec
specific class type is assigned - A2dpCodecConfigExt. Since it's
assigned only to provider supported codecs, we can assume if a codec is
supported using this info rather then call hal dependent
provider::supports_codec function.

Bug: 374984535
Flag: EXEMPT - no logical change
Test: m com.google.android.bt
Change-Id: I9f3a44b3bd09fd7c246d9e419db059389589c82a