connection_manager: Fix handling direct connect timeout
When timeout happens for direct connect, connection manager should
clear its state in first place and then notify others about direct connect being
timeout.
This is important to handle it internally first, as connection manager
can take a decision to get device back to accept list if there is a
client which expects this.

Recently due to:

157d12ef10 Conclude direct connection requests on connection failure

when upper layer got notified about timeout and immediately called
connection_manager::on_connection_complete() function for some reason
which cause clearing the device in the way that later it could not be
added to accept list, even it should.

Issue happens if following scenario
1. Disable Bluetooth
2. Take 1 bud out of the case and enable Bluetooth
3. Wait 30 sec
4. Take 2 bud out of the case -> 2 bud will not reconnect as it is not
   in accept list.

This CL fixes that.

Bug: 401435350
Test: atest net_test_gatt_conn_multiplexing
Flag: EXEMPT, regression handled with unit test, new test added

Change-Id: I66ff0e36642c5b3997a3990857b2585e9718cddc