Correct BluetoothPbapUtils#fetchPbapParams
In the `fetchPbapParams` method, the incorrect variable
was being used to retrieve the "totalContacts" value.

This resulted in the `sTotalContacts` variable not being populated as expected.

This commit corrects the variable usage,
ensuring that "totalContacts" is fetched properly.

Additionally, a new test has been added to verify the fix and prevent regressions.

Bug: 410481988
Flag: EXEMPT, trivial change
Test: atest BluetoothPbapUtilsTest
Change-Id: Idfe4341122475311e513a13cd0ff203b76de7607