LeAudio: Adjust GAME vs CONVERSATIONAL policy
With this change, if GMAP is available on the remote device,
it will be prioritized even during a Call or VoIP usage.
This will allow using voice communication when in a GMAP
mode instead of switching to CONVERSATIONAL use case.

Bug: 406589272
Bug: 398764690
Test: atest bluetooth_le_audio_client_test
Flag: com.android.bluetooth.flags.leaudio_use_context_type_manager
Change-Id: I1255e17bae6f8de063cb5da43a72f51f418e7a1d