List connections by ID and device for clients and servers
Problem: We've been seeing bugs when connecting to devices using than
one bearer at a time. We lack the easy bugging to see when this is the
case. We have the goal to change up some of this logic, so further
debugging will be key to development and stabilizing here.

Solution: Print the list of connections for all clients and servers.
There is a map for clients and servers which contains all the app recors
and connection records that come from the native stack. Underneath,
these connection IDs essentially map to a single bearer in the native
stack. We currently only support one bearer per transport when
connecting with a device. Each device can have up to two connections per
app in this list.

Flag: EXEMPT, logging only
Bug: 382778345
Test: atest BluetoothInstrumentationTest
Change-Id: I035fae592b43a82f4f493e6965b32b62571987d7