Send transport with GATT client and server connection events
Problem: In java, we receive connection lifecycle events and use them to
build a cache of connection state. Other code uses this cache to make
decisions like which connection ID to target for a given device. Since
finding bugs when dealing with devices that have multiple ATT
bearers/underlying connections to the remove device, we've decided to
introduce a function to get _all_ connection IDs for a given device.
Now, when code goes to pick a connection ID for a device, they don't
have any useful context to decide which one to use. We need as much
context as we can get so we pick the desired connection ID to make
operations with.

Solution: Since we only have one bearer per transport, the decision
right now reduces to picking either BR/EDR or LE. Both clients and
servers have transport preferences that can be set when creating them.
We can pipe transport from the connection and disconnection events and
add that context to our ContextMap. Now, the transport can be used to
target the preferred transport of a client or server.

Flag: EXEMPT, mechanical refactor
Bug: 382778345
Test: m com.android.bt
Change-Id: I794b09e78c0820d2cc0f40db08095592d01dfca9