gatt: Move callouts to after null checks
During shutdown, callbacks can get posted on the JNI thread whilst
shutdown takes place on a different thread.

To eliminate the possibility of calling into the Rust private Gatt
server when it has been shut down, this CL performs the shut down checks
first.

A consequence of this is that a lock will be held whilst calling the
Rust code, which could have the potential to cause deadlocks.  This
should not be an issue because the Rust code handles all callouts
asynchronously.

Bug: 411876748
Flag: EXEMPT bug fix
Test: m com.android.btservices
Change-Id: I7e312534a75c30b5bf245ec07ec126fc567afaca