leaudio: Expose helper function to set configuration_context_type
Configuration_context_type_ represents current major scenario defing ISO
configuration for the stream.
This is calculated based on the metadata set by Audio HAL and depends on
Supported and Available context types on the remote side.

This CL creates helper function to set this variable and log it for
debugging purposes.

Bug: 331775328
Test: atest bluetooth_le_audio_client_test
Flag: EXEMPT, mechanical refactor

Change-Id: I7e1d65ad85e88ccda5f810196e356c39f69cbf85