HIDH: Wait for HID disconnection before marking a device unused
When removing a device, we immediately mark it as unused even though
the disconnection process is still ongoing. Because we need to
disconnect the HID channels one by one, we might be stuck in an
intermediate step, e.g. interrupt channel is disconnected but the
control channel is not, because we can't retrieve the device info
anymore. Then depending on the peer device, they might or might not
continue disconnecting the control channel.

Here we put the action to mark it unused as pending when we remove a
device. It will be carried out once all HID channels are disconnected.
This is achieved by introducing a new state "REMOVING".

Additionally, this CL also consolidate the usage of in_use and the
state itself.

Bug: 406577215
Bug: 412876005
Test: m -j
Flag: com.android.bluetooth.flags.wait_hid_disconnect_before_marking_unused

Change-Id: Ib92d9a20c32fa5fce2e7bcf38776c2790098c99d