SystemServer: Fix tesst with death_recipient tests
In order to make test faster, the mBinder is posting is death
immediately. This is call during disable.
But enableBle is monitoring lifecycle and this binder should not die.

Also, move some assertion to the end of each test vs the teardown, to
make sure the error message in ath is correctly represented

Bug: 414678172
Test: atest ServiceBluetoothTests
Flag: com.android.bluetooth.flags.ble_death_recipient_thread
Change-Id: I6c7ab2161d7179daa13a07d39a4b8a9fa2c78f35