Identity address of bonded must be always available
Identity addresses of the bonded devices are not set on BT restart if they are the same as the pseudo addresses. When identity addresses of the bonded devices are not available, it must be assumed to be the same as the pseudo address.

Test: mmm packages/modules/Bluetooth
Flag: com.android.bluetooth.flags.always_set_identity_addr
Bug: 390395989
Bug: 415404383
Change-Id: I840cb11a1d553fb3b69ce37daa9a2e66b85d7160