Use `TestUtils#getRealDevice()` in all* unit tests
*all = unfortunately there are still some tests that depend
on a static `AdapterService`. Those are not able to be handled
at the moment until we can relieve them of their static getter
access and replace with a mock via:
`when(mAdapterService.getRemoteDevice(anyString()))`

This is essentially 100% no-op and the same thing as calling
`BluetoothAdapter.getDefaultAdapter().getRemoteDevice(address);`

This is being done in preparation of the potential support
for Ravenwood tests.

Bug: 413308573
Test: atest AvrcpControllerNativeInterfaceTest
Test: atest CsipSetCoordinatorServiceTest
Test: atest DatabaseManagerTest
Test: atest GattServiceTest
Test: atest HearingAidNativeInterfaceTest
Test: atest HfpClientCallTest
Test: atest LeAudioBroadcasterNativeInterfaceTest
Test: atest LeAudioNativeInterfaceTest
Test: atest PhonePolicyTest
Test: atest ScanControllerTest
Flag: EXEMPT test only
Change-Id: I3ea50808636d33f1113461e3b1ec9817c96730df