leaudio: Fix setting group to active while phone call
When Group is setting to active while phone call is ongoing,
the Group is put to QoS Configured state. Later on resume,
when state machine is asking for enabled direction, the audio_sender_state_
or/and audio_receiver_state_ are not yet set, which makes it impossible
to move to Enabling state.
This CL fixes that.

Bug: 380807580
Bug: 412540131
Bug: 415811749
Test: atest bluetooth_le_audio_client_test
Flag: com.android.bluetooth.flags.leaudio_dynamic_direction_opening
Change-Id: I6de496b9982f7bfdc16e2c8166c2ae54b7652f1e